better handling of event unsubscription
You no longer have to manually unsubscribe from events when an argument is destroyed.

EventSubscriber now keeps a list of all event types it's tracking. This list is added to whenever we invoke SubscribeToList; when an argument is destroyed, we iterate through the list and unsubscribe the argument from each event type it was tracking.

TL;DR: Manual subscription to events is necessary for arguments (as expected), but unsubscription is now automatic when the argument is destroyed.