Hopeful fix to action resolution system
This fixes a bug found by DeckardAddendum.

The old AddAction() always tries to empty the action queue once an action is added to the queue. However, this causes an indexing issue if we add actions to an already-existing queue clear process.

The current fix is to check to see if something is already emptying the queue, and if so, don't try to empty the queue again (since that would repeat the action currently being handled by the action queue!).

AddAction() will always add the listed action to the end of the queue, so the original queue clear process should catch the newly-added action.